addon: move addon subcommands to subpackages - verbatim
