+++ cmd/wksctl/addon/list/list.go
@@ -1,4 +1,4 @@
-package main
+package list
 
 import (
 	"fmt"
